+++ mac/Sources/GhostComplete/AutocompletePreferences.swift
@@ -7,14 +7,32 @@ struct AutocompletePreferences: Codable, Equatable, Sendable {
     var revealStepMs: Int
     var overlayNudgeX: Int
     var overlayNudgeY: Int
+    var rawTextLoggingEnabled: Bool
 
     static let `default` = AutocompletePreferences(
         debounceMs: 120,
         revealAnimationEnabled: true,
         revealStepMs: 18,
-        overlayNudgeX: 1,
-        overlayNudgeY: 0
+        overlayNudgeX: 0,
+        overlayNudgeY: 0,
+        rawTextLoggingEnabled: false
     )
+
+    init(
+        debounceMs: Int,
+        revealAnimationEnabled: Bool,
+        revealStepMs: Int,
+        overlayNudgeX: Int,
+        overlayNudgeY: Int,
+        rawTextLoggingEnabled: Bool = false
+    ) {
+        self.debounceMs = debounceMs
+        self.revealAnimationEnabled = revealAnimationEnabled
+        self.revealStepMs = revealStepMs
+        self.overlayNudgeX = overlayNudgeX
+        self.overlayNudgeY = overlayNudgeY
+        self.rawTextLoggingEnabled = rawTextLoggingEnabled
+    }
 
     var debounceDelay: TimeInterval {
         TimeInterval(clamped(debounceMs, min: 60, max: 600)) / 1000
@@ -38,8 +56,28 @@ struct AutocompletePreferences: Codable, Equatable, Sendable {
             revealAnimationEnabled: revealAnimationEnabled,
             revealStepMs: clamped(revealStepMs, min: 5, max: 120),
             overlayNudgeX: clamped(overlayNudgeX, min: -40, max: 40),
-            overlayNudgeY: clamped(overlayNudgeY, min: -40, max: 40)
+            overlayNudgeY: clamped(overlayNudgeY, min: -40, max: 40),
+            rawTextLoggingEnabled: rawTextLoggingEnabled
         )
+    }
+
+    private enum CodingKeys: String, CodingKey {
+        case debounceMs
+        case revealAnimationEnabled
+        case revealStepMs
+        case overlayNudgeX
+        case overlayNudgeY
+        case rawTextLoggingEnabled
+    }
+
+    init(from decoder: Decoder) throws {
+        let container = try decoder.container(keyedBy: CodingKeys.self)
+        debounceMs = try container.decode(Int.self, forKey: .debounceMs)
+        revealAnimationEnabled = try container.decode(Bool.self, forKey: .revealAnimationEnabled)
+        revealStepMs = try container.decode(Int.self, forKey: .revealStepMs)
+        overlayNudgeX = try container.decode(Int.self, forKey: .overlayNudgeX)
+        overlayNudgeY = try container.decode(Int.self, forKey: .overlayNudgeY)
+        rawTextLoggingEnabled = try container.decodeIfPresent(Bool.self, forKey: .rawTextLoggingEnabled) ?? false
     }
 }